EAD-Session-10

+++ EAD/employeePortal/src/main/java/com/metacube/training/dao/ProjectDAO.java
@@ -7,7 +7,7 @@ import com.metacube.training.model.Project;
 public interface ProjectDAO {
 	
 	
-	Project getProjectById(Long id);
+	Project getProjectById(int id);
 
 	List<Project> getAllProjects();
