+++ EAD/employeePortal/src/main/java/com/metacube/training/service/ProjectServiceImpl.java
@@ -16,7 +16,7 @@ public class ProjectServiceImpl implements ProjectService {
 	
 	
 	@Override
-	public Project getProjectById(Long id) {
+	public Project getProjectById(int id) {
 		
 		return projectDAO.getProjectById(id);
 	}
@@ -27,7 +27,7 @@ public class ProjectServiceImpl implements ProjectService {
 	}
 
 	@Override
-	public boolean deleteProject(Long id) {
+	public boolean deleteProject(int id) {
 		Project project = projectDAO.getProjectById(id);
 		return projectDAO.deleteProject(project);
 	}
